cost_tracker_cloud_function/requirements.txt
functions-framework
google-cloud-bigquery
openai